backend/requirements.txt
# Flask and extensions
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10

# Database
SQLAlchemy==2.0.7
PyMySQL==1.0.3
cryptography==39.0.2

# Utilities
python-dotenv==1.0.0
Faker==18.3.1
Werkzeug==2.2.3

# Testing
pytest==7.3.1
